Update flat page and panel tests
