+++ tests/flat_snapshots/page_with_title.json
@@ -1,0 +1,76 @@
+[
+	{
+		"brickType": "CompositeBrick",
+		"width": 538.5826771653543,
+		"height": 44.849999999999994,
+		"xLeft": 28.34645669291339,
+		"xRight": 566.9291338582677,
+		"yTop": 28.34645669291339,
+		"yBot": 73.19645669291339,
+		"bricks": [
+			{
+				"brickType": "CompositeBrick",
+				"width": 152.88,
+				"height": 44.849999999999994,
+				"xLeft": 28.34645669291339,
+				"xRight": 181.2264566929134,
+				"yTop": 28.34645669291339,
+				"yBot": 73.19645669291339,
+				"bricks": [
+					{
+						"brickType": "TextBrick",
+						"width": 152.88,
+						"height": 44.849999999999994,
+						"xLeft": 28.34645669291339,
+						"xRight": 181.2264566929134,
+						"yTop": 28.34645669291339,
+						"yBot": 73.19645669291339,
+						"text": "So Page",
+						"options": {
+							"fontSize": 39,
+							"fontName": "helvetica",
+							"fontStyle": "bold",
+							"fontColor": "#404040"
+						}
+					}
+				]
+			},
+			{
+				"brickType": "RowlineBrick",
+				"width": 538.5826771653543,
+				"height": 0,
+				"xLeft": 28.34645669291339,
+				"xRight": 566.9291338582677,
+				"yTop": 73.19645669291339,
+				"yBot": 73.19645669291339
+			}
+		]
+	},
+	{
+		"brickType": "TextFieldBrick",
+		"width": 504.0826771653543,
+		"height": 34.5,
+		"xLeft": 62.84645669291339,
+		"xRight": 566.9291338582677,
+		"yTop": 107.69645669291339,
+		"yBot": 142.1964566929134,
+		"options": {
+			"fieldName": "sq_104",
+			"inputType": "text",
+			"value": "",
+			"isReadOnly": false,
+			"shouldRenderBorders": true,
+			"placeholder": "",
+			"isMultiline": false
+		}
+	},
+	{
+		"brickType": "RowlineBrick",
+		"width": 538.5826771653543,
+		"height": 0,
+		"xLeft": 28.34645669291339,
+		"xRight": 566.9291338582677,
+		"yTop": 193.9464566929134,
+		"yBot": 193.9464566929134
+	}
+]
