+++ tests/flat_snapshots/panel_inner_indent.json
@@ -1,0 +1,87 @@
+[
+	{
+		"brickType": "CompositeBrick",
+		"width": 435.0826771653543,
+		"height": 89.69999999999999,
+		"xLeft": 131.8464566929134,
+		"xRight": 566.9291338582677,
+		"yTop": 28.34645669291339,
+		"yBot": 118.04645669291338,
+		"bricks": [
+			{
+				"brickType": "CompositeBrick",
+				"width": 258.05999999999995,
+				"height": 37.94999999999999,
+				"xLeft": 131.8464566929134,
+				"xRight": 389.90645669291337,
+				"yTop": 28.34645669291339,
+				"yBot": 66.29645669291338,
+				"bricks": [
+					{
+						"brickType": "CompositeBrick",
+						"width": 258.05999999999995,
+						"height": 37.94999999999999,
+						"xLeft": 131.8464566929134,
+						"xRight": 389.90645669291337,
+						"yTop": 28.34645669291339,
+						"yBot": 66.29645669291338,
+						"bricks": [
+							{
+								"brickType": "TextBrick",
+								"width": 258.05999999999995,
+								"height": 37.94999999999999,
+								"xLeft": 131.8464566929134,
+								"xRight": 389.90645669291337,
+								"yTop": 28.34645669291339,
+								"yBot": 66.29645669291338,
+								"text": "I am in the panel",
+								"options": {
+									"fontSize": 33,
+									"fontName": "helvetica",
+									"fontStyle": "bold",
+									"fontColor": "#404040"
+								}
+							}
+						]
+					}
+				]
+			},
+			{
+				"brickType": "RowlineBrick",
+				"width": 435.0826771653543,
+				"height": 0,
+				"xLeft": 131.8464566929134,
+				"xRight": 566.9291338582677,
+				"yTop": 66.29645669291338,
+				"yBot": 66.29645669291338
+			},
+			{
+				"brickType": "TextFieldBrick",
+				"width": 400.5826771653543,
+				"height": 34.5,
+				"xLeft": 166.3464566929134,
+				"xRight": 566.9291338582677,
+				"yTop": 83.54645669291338,
+				"yBot": 118.04645669291338,
+				"options": {
+					"fieldName": "sq_104",
+					"inputType": "text",
+					"value": "",
+					"isReadOnly": false,
+					"shouldRenderBorders": true,
+					"placeholder": "",
+					"isMultiline": false
+				}
+			}
+		]
+	},
+	{
+		"brickType": "RowlineBrick",
+		"width": 435.0826771653543,
+		"height": 0,
+		"xLeft": 131.8464566929134,
+		"xRight": 566.9291338582677,
+		"yTop": 169.79645669291338,
+		"yBot": 169.79645669291338
+	}
+]
