+++ tests/flat_snapshots/panel_with_invisible_questions.json
@@ -1,0 +1,145 @@
+[
+	{
+		"brickType": "CompositeBrick",
+		"width": 538.5826771653543,
+		"height": 44.849999999999994,
+		"xLeft": 28.34645669291339,
+		"xRight": 566.9291338582677,
+		"yTop": 28.34645669291339,
+		"yBot": 73.19645669291339,
+		"bricks": [
+			{
+				"brickType": "CompositeBrick",
+				"width": 194.22000000000003,
+				"height": 44.849999999999994,
+				"xLeft": 28.34645669291339,
+				"xRight": 222.56645669291342,
+				"yTop": 28.34645669291339,
+				"yBot": 73.19645669291339,
+				"bricks": [
+					{
+						"brickType": "CompositeBrick",
+						"width": 194.22000000000003,
+						"height": 44.849999999999994,
+						"xLeft": 28.34645669291339,
+						"xRight": 222.56645669291342,
+						"yTop": 28.34645669291339,
+						"yBot": 73.19645669291339,
+						"bricks": [
+							{
+								"brickType": "TextBrick",
+								"width": 194.22000000000003,
+								"height": 44.849999999999994,
+								"xLeft": 28.34645669291339,
+								"xRight": 222.56645669291342,
+								"yTop": 28.34645669291339,
+								"yBot": 73.19645669291339,
+								"text": "Panel Title",
+								"options": {
+									"fontSize": 39,
+									"fontName": "helvetica",
+									"fontStyle": "bold",
+									"fontColor": "#404040"
+								}
+							}
+						]
+					}
+				]
+			},
+			{
+				"brickType": "RowlineBrick",
+				"width": 538.5826771653543,
+				"height": 0,
+				"xLeft": 28.34645669291339,
+				"xRight": 566.9291338582677,
+				"yTop": 73.19645669291339,
+				"yBot": 73.19645669291339
+			}
+		]
+	},
+	{
+		"brickType": "CompositeBrick",
+		"width": 538.5826771653543,
+		"height": 89.69999999999999,
+		"xLeft": 28.34645669291339,
+		"xRight": 566.9291338582677,
+		"yTop": 107.69645669291339,
+		"yBot": 197.39645669291338,
+		"bricks": [
+			{
+				"brickType": "CompositeBrick",
+				"width": 258.06,
+				"height": 37.94999999999999,
+				"xLeft": 28.34645669291339,
+				"xRight": 286.40645669291337,
+				"yTop": 107.69645669291339,
+				"yBot": 145.64645669291338,
+				"bricks": [
+					{
+						"brickType": "CompositeBrick",
+						"width": 258.06,
+						"height": 37.94999999999999,
+						"xLeft": 28.34645669291339,
+						"xRight": 286.40645669291337,
+						"yTop": 107.69645669291339,
+						"yBot": 145.64645669291338,
+						"bricks": [
+							{
+								"brickType": "TextBrick",
+								"width": 258.06,
+								"height": 37.94999999999999,
+								"xLeft": 28.34645669291339,
+								"xRight": 286.40645669291337,
+								"yTop": 107.69645669291339,
+								"yBot": 145.64645669291338,
+								"text": "I am in the panel",
+								"options": {
+									"fontSize": 33,
+									"fontName": "helvetica",
+									"fontStyle": "bold",
+									"fontColor": "#404040"
+								}
+							}
+						]
+					}
+				]
+			},
+			{
+				"brickType": "RowlineBrick",
+				"width": 538.5826771653543,
+				"height": 0,
+				"xLeft": 28.34645669291339,
+				"xRight": 566.9291338582677,
+				"yTop": 145.64645669291338,
+				"yBot": 145.64645669291338
+			},
+			{
+				"brickType": "TextFieldBrick",
+				"width": 504.0826771653543,
+				"height": 34.5,
+				"xLeft": 62.84645669291339,
+				"xRight": 566.9291338582677,
+				"yTop": 162.89645669291338,
+				"yBot": 197.39645669291338,
+				"options": {
+					"fieldName": "sq_106",
+					"inputType": "text",
+					"value": "",
+					"isReadOnly": false,
+					"shouldRenderBorders": true,
+					"placeholder": "",
+					"isMultiline": false
+				}
+			}
+		]
+	},
+	{
+		"brickType": "RowlineBrick",
+		"width": 538.5826771653543,
+		"height": 0,
+		"xLeft": 28.34645669291339,
+		"xRight": 566.9291338582677,
+		"yTop": 249.14645669291338,
+		"yBot": 249.14645669291338
+	}
+]
